+++ src/shared/types/option.ts
@@ -5,8 +5,7 @@ export type Option<T = string, L = React.ReactNode> = {
 };
 
 export type OptionGroup<T> = {
-  label: string;
-  displayLabel?: string;
+  label?: string;
   options: Option<T>[];
 };